A bit of refactoring before adding tests
