+++ x/tokenomics/keeper/keeper_update_relay_mining_difficulty_test.go
@@ -1,0 +1,14 @@
+package keeper_test
+
+func (s *TestSuite) UpdateRelayMiningDifficulty_First() {
+	// Retrieve default values
+	// t := s.T()
+	// ctx := s.ctx
+	// sdkCtx := sdk.UnwrapSDKContext(ctx)
+
+	// // 0. Add the claim & verify it exists
+	// claim := s.claim
+	// s.keepers.UpsertClaim(ctx, claim)
+	// claims := s.keepers.GetAllClaims(ctx)
+	// s.Require().Len(claims, 1)
+}
